Converted all long values to decimal

+++ src/backend/MoneySpot6.WebApp/Database/Db.cs
@@ -52,7 +52,7 @@ namespace MoneySpot6.WebApp.Database
         public required string CustomerId { get; set; }
         public required string AccountType { get; set; }
         public required string Type { get; set; }
-        public required long Balance { get; set; }
+        public required decimal Balance { get; set; }
 
     }
 
@@ -73,10 +73,10 @@ namespace MoneySpot6.WebApp.Database
         public required CounterpartyAccount Counterparty { get; set; }
         public string? Purpose { get; set; }
         public string? Code { get; set; }
-        public long Amount { get; set; }
-        public long? OriginalAmount { get; set; }
-        public long? ChargeAmount { get; set; }
-        public long NewBalance { get; set; }
+        public decimal Amount { get; set; }
+        public decimal? OriginalAmount { get; set; }
+        public decimal? ChargeAmount { get; set; }
+        public decimal NewBalance { get; set; }
         public bool IsCancelation { get; set; }
         public string? CustomerReference { get; set; }
         public string? InstituteReference { get; set; }
